Extract E2E test for registry
